+++ R/expand.R
@@ -3,35 +3,38 @@
 #' Any tidy data frame of matrices (in which each row represents one matrix observation)
 #' can also be represented as a tidy data frame
 #' with each non-zero matrix entry as an observation on its own row.
-#' This function (and `collapse_to_matrices()`) convert between the two representations.
+#' This function and `collapse_to_matrices()` convert between the two representations.
 #'
 #' Names for output columns are specified in the `rownames`, `colnames`,
 #' `rowtypes`, and `coltypes`, arguments.
 #' The entries of the \pkg{matsindf}-style matrices are stored in an output column named `values.`
 #'
-#' @param      .DF The data frame containing \pkg{matsindf}-style matrices.
-#'                 (`.DF` may also be a named list of matrices, in which case
-#'                 names of the matrices are taken from the names of items in the list and
-#'                 list items are expected to be matrices.)
+#' @param .DF The data frame containing \pkg{matsindf}-style matrices.
+#'            (`.DF` may also be a named list of matrices, in which case
+#'            names of the matrices are taken from the names of items in the list and
+#'            list items are expected to be matrices.)
 #' @param matnames The name of the column in `.DF` containing matrix names (a string).
 #'                 Default is "matnames".
-#' @param  matvals The name of the column in `.DF` containing IO-style matrices or constants (a string),
-#'                 This will also be the name of the column containing matrix entries in the output data frame.
-#'                 Default is "matvals".
+#' @param matvals The name of the column in `.DF` containing IO-style matrices
+#'                or constants (a string).
+#'                This will also be the name of the column containing matrix entries
+#'                in the output data frame.
+#'                Default is "matvals".
 #' @param rownames The name for the output column of row names (a string).
 #'                 Default is "rownames".
 #' @param colnames The name for the output column of column names (a string).
 #'                 Default is "colnames".
 #' @param rowtypes An optional name for the output column of row types (a string).
 #'                 Default is "rowtypes".
-#' @param coltypes The optional name for the output column of column types (a string)
+#' @param coltypes The optional name for the output column of column types (a string).
 #'                 Default is "coltypes".
-#' @param     drop If specified, the value to be dropped from output,
-#'                 For example, `drop = 0` will cause `0` entries in the matrices to be deleted from output.
-#'                 If `NA`, no values are dropped from output.
-#'                 Default is `NA`.
+#' @param drop If specified, the value to be dropped from output,
+#'             For example, `drop = 0` will cause `0` entries in the matrices
+#'             to be deleted from output.
+#'             If `NA`, no values are dropped from output.
+#'             Default is `NA`.
 #'
-#' @return A tidy data frame containing expanded \pkg{matsindf}-style matrices
+#' @return A tidy data frame containing expanded \pkg{matsindf}-style matrices.
 #'
 #' @export
 #'
@@ -67,9 +70,13 @@
 #' expand_to_tidy(mats, matnames = "matrix", matvals = "vals",
 #'                      rownames = "rows", colnames = "cols",
 #'                      rowtypes = "rt",   coltypes = "ct", drop = 0)
-expand_to_tidy <- function(.DF, matnames = "matnames", matvals = "matvals",
-                           rownames = "rownames", colnames = "colnames",
-                           rowtypes = "rowtypes", coltypes = "coltypes",
+expand_to_tidy <- function(.DF,
+                           matnames = "matnames",
+                           matvals = "matvals",
+                           rownames = "rownames",
+                           colnames = "colnames",
+                           rowtypes = "rowtypes",
+                           coltypes = "coltypes",
                            drop = NA){
   if (!is.data.frame(.DF) & is.list(.DF)) {
     # Create an empty 1-row data frame with row names taken from .DF and promote to a column
@@ -80,15 +87,26 @@ expand_to_tidy <- function(.DF, matnames = "matnames", matvals = "matvals",
     tempDF[[matvals]] <- I(.DF)
     .DF <- tempDF
   }
-  .DF %>%
+
+  temp <- .DF |>
+    # Remove any rows with NULL entries in the matvals column.
+    dplyr::filter(! sapply(.data[[matvals]], is.null))
+  if (nrow(temp) == 0) {
+    return(NULL)
+  }
+  temp |>
     # group by everything except matvals column so that "do" will act as desired
-    dplyr::group_by_at(setdiff(colnames(.DF), matvals)) %>%
+    dplyr::group_by_at(setdiff(colnames(.DF), matvals)) |>
     dplyr::do(
       # Convert .data to row, col, val format
-      mat_to_rowcolval(.data[[matvals]][[1L]], rownames = rownames, colnames = colnames,
-                       rowtypes = rowtypes, coltypes = coltypes,
-                       matvals = matvals, drop = drop)
-    ) %>%
+      mat_to_rowcolval(.data[[matvals]][[1L]],
+                       rownames = rownames,
+                       colnames = colnames,
+                       rowtypes = rowtypes,
+                       coltypes = coltypes,
+                       matvals = matvals,
+                       drop = drop)
+    ) |>
     # Remove the grouping
     dplyr::ungroup()
 }